Ficha4

Description

Este módulo contém definições Haskell para o cálculo de funções recursivas simples.

Documentation

mydiv :: Float -> Float -> Float

Funcao que recebe dois numeros Float e devolve a sua divisao.

Exemplos de utilização:

Expand

>>> mydiv 4 2
2.0
>>> mydiv 6 2
3.0
>>> mydiv 3 2
1.5

mydiv1 :: Int -> Int -> Int

Funcao que recebe dois numeros Int e devolve a sua divisao.

Exemplos de utilização:

Expand

>>> mydiv1 4 2
2
>>> mydiv1 6 2
3
>>> mydiv1 3 2
1

type Matriz a

Arguments

= [[a]]
Definicao de um tipo Matriz

matrizExemplo1 :: Matriz Int

Uma matriz exemplo [[1, 2, 3], [4, 5, 6], [7, 8, 9]]

matrizExemplo2 :: Matriz Int

Uma matriz exemplo [[1, 2], [3, 4]]

trocaPrimLinhaUlt :: Matriz a -> Matriz a

Funcao que recebe uma matriz e troca a primeira linha com a ultima.

Exemplos de utilização:

Expand

>>> [[1, 2, 3], [4, 5, 6], [7, 8, 9]]
[[7,8,9],[4,5,6],[1,2,3]]

>>> [[1, 2], [3, 4]]
[3, 4], [1, 2]]